updated supercell texture tool to support sc2

+++ supercell-flash/source/flash/objects/SupercellSWF2.h
@@ -37,7 +37,7 @@ namespace sc::flash
 			auto it = std::find(std::execution::par_unseq, strings.begin(), strings.end(), target);
 			if (it != strings.end())
 			{
-				return std::distance(strings.begin(), it);
+				return (uint32_t)std::distance(strings.begin(), it);
 			}
 			else
 			{
